Correct some issues with reproducer code/logic
Running the reproducer against a multi-compute scenario raised an issue
in the established logic in said step.

We now wait for SSH using the proper IP, and from the hypervisor instead
of some fancy localhost that wasn't configured until... next step.

We also take the opportunity to leverage a better filter for IP address
extraction.

This patch also corrects how repositories were cloned, especially when it
comes to github repository and the tested PR. Since most of the
repositories work with forks instead of dev-branches, the PR code/change
isn't usually known by the main repository - and we have to switch to
some specific reference instead.

We also take the opportunity to ensure the controller-0 knows the public
ssh key of all the hosts, even itself.