Correctly handle types in kustomize_compute template
Without this patch, a 3 node edpm-compute deployment fails
with "can only concatenate str (not \"int\") to str".

Update the kustomize_compute.yml.j2 template in the
libvirt_manager role to explicitly cast the item to
an int so that addition can take place and then cast
the result to a string so that concatenation can take
place.

Signed-off-by: John Fulton <fulton@redhat.com>