Replace copy with slurp in dlrn_promote check criteria tasks
Currently the dlrn_promote jobs are failing with following error:
```
TASK [dlrn_promote : Load criteria for promotion]
2023-10-16 04:02:33.078956 | primary | ERROR
2023-10-16 04:02:33.079200 | primary | {
2023-10-16 04:02:33.079236 | primary |   "ansible_facts": {
2023-10-16 04:02:33.079262 | primary |     "criteria": {}
2023-10-16 04:02:33.079286 | primary |   },
2023-10-16 04:02:33.079308 | primary |   "ansible_included_var_files": [],
2023-10-16 04:02:33.079348 | primary |   "message": "Could not find or access '/home/zuul/ci-framework-data/antelope.yaml' on the Ansible Controller.\nIf you are using a module and expect the file to exist on the remote, see the remote_src option"
```

The Criteria file exists in the workspace. But it is failing to load.
In order to fix the issue. We can slurp the criteria file and store
the needed data in criteria var. It will fix the issue.

This pr adds molecule tests for check_criteria taskfile.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>